Update the base class to be independent from bq
